+++ server/src/main/java/org/opensearch/action/ActionScope.java
@@ -12,13 +12,18 @@ import org.opensearch.identity.scopes.ScopeEnums.ScopeNamespace;
 import org.opensearch.identity.scopes.ScopeEnums.ScopeArea;
 
 /**
- * Scopes associated with actions in OpenSearch
+ * This enum defines the ActionScope Scope implementation.
+ *
+ * An ActionScope is composed of a ScopeNamespace, ScopeArea, and Action.
+ * Together, these three components reference a Scope which confers execution privileges to a Subject. For example, the READ ActionScope
+ * is has a ScopeArea of INDEX and an Action of READ. Its ScopeNamespace is ACTION since it is an ActionScope. This ActionScope is required by all
+ * Subjects which seek to perform read-related operations on an INDEX.
  *
  * @opensearch.experimental
  */
 public enum ActionScope implements Scope {
     ALL(ScopeArea.ALL, "ALL"),
-    READ(ScopeArea.ALL, "READ");
+    READ(ScopeArea.INDEX, "READ");
 
     public final ScopeArea area;
     public final String action;