+++ server/src/main/java/org/opensearch/identity/ApplicationScope.java
@@ -10,13 +10,17 @@ import org.opensearch.identity.scopes.ScopeEnums.ScopeArea;
 import org.opensearch.identity.scopes.ScopeEnums.ScopeNamespace;
 
 /**
- * Generic Scopes in OpenSearch
+ * This enum defines the ApplicationScope implementation of the Scope interface. Like other Scope implementations, ApplicationScope enum constants
+ * contain a ScopeNamespace, ScopeArea, and Action.
+ *
+ * An ApplicationScope dictates special privileges an application may have. For example, the SuperUserAccess ApplicationScope provides an application
+ * trust equal to that of an administrator.
  *
  * @opensearch.experimental
  */
 public enum ApplicationScope implements Scope {
-    TRUSTED(ScopeArea.ALL, "TRUSTED"),
-    UNTRUSTED(ScopeArea.ALL, "UNTRUSTED");
+    SuperUserAccess(ScopeArea.ALL, "ALL");
+
 
     public final ScopeArea area;
     public final String action;