+++ CHANGELOG.md
@@ -95,6 +95,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add distributed tracing framework ([#7543](https://github.com/opensearch-project/OpenSearch/issues/7543))
 - Enable Point based optimization for custom comparators ([#8168](https://github.com/opensearch-project/OpenSearch/pull/8168))
 - Introduce Limited Scopes for Extension Management ([#7850]https://github.com/opensearch-project/OpenSearch/pull/7850)
+- Introduce Application Scopes for Extension Management ([#7850]https://github.com/opensearch-project/OpenSearch/pull/7850)
+
 
 ### Dependencies
 - Bump `com.azure:azure-storage-common` from 12.21.0 to 12.21.1 (#7566, #7814)